02 - Operators
Given the meal price (base cost of a meal), tip percent (the percentage
of the meal price being added as tip), and tax percent (the percentage
of the meal price being added as tax) for a meal, find and print the
meal's total cost. Round the result to the nearest integer.

Complete the solve function in the editor below.solve has the following
parameters:

int meal_cost: the cost of food before tip and tax
int tip_percent: the tip percentage
int tax_percent: the tax percentage

The function returns nothing. Print the calculated value, rounded to the
nearest integer.