11 - 2D Array
Given a 6x5 2D Array, A:

1 1 1 0 0 0
0 1 0 0 0 0
1 1 1 0 0 0
0 0 0 0 0 0
0 0 0 0 0 0
0 0 0 0 0 0

We define an hourglass in A to be a subset of values with indices
falling in this pattern in A's graphical representation:

a b c
  d
e f g

There are 16 hourglasses in A, and an hourglass sum is the sum of an
hourglass' values.

Calculate the hourglass sum for every hourglass in A, then print the
maximum hourglass sum.

Constraints:

-9 <= A[i][j] <= 9
0 <= i,j <= 5