22 - Binary Search Tree: Height
The height of a binary search tree is the number of edges between the
tree's root and its furthest leaf. You are given a pointer, root,
pointing to the root of a binary search tree. Complete the getHeight
function provided in your editor so that it returns the height of the
binary search tree.

Input Format:

The locked stub code in your editor reads the following inputs and
assembles them into a binary search tree:

- The first line contains an integer, n, denoting the number of nodes in
the tree.
- Each of the subsequent lines contains an integer,n, denoting the value
of an element that must be added to the BST.

Output Format:

The locked stub code in your editor will print the integer returned by
your getHeight function denoting the height of the BST.