25 - Running Time and Complexity
A prime is a natural number greater than 1 that has no positive divisors
other than and itself. Given a number, n, determine and print whether it
is Prime or Not prime.

Note: If possible, try to come up with a O(√n) primality algorithm, or
see what sort of optimizations you come up with for an O(n) algorithm.