27 - Testing
A coworker has prepared functions that will perform the tests and
validate return values. Finish the implementation of 3 classes to
provide data and expected results for the tests.

Complete the following methods.

1. In the class TestDataEmptyArray:

get_array() returns an empty array

2. In the class TestDataUniqueValues:

- get_array() returns an array of size at least 2 with all unique
elements

- get_expected_result() returns the expected minimum value index for
this array

3. In the class TestDataExactlyTwoDifferentMinimums:

- get_array() returns an array where the minimum value occurs at exactly
2 indices

- get_expected_result() returns the expected index

Take a look at the code template to see the exact implementation of
functions that your colleague already implemented.